Memperbaiki bug pada bagian tambah, edit, hapus gambar agar gambar yang mengalami perubahan atau dihapus maka file tersebut akan hilang pada penyimpanan serta memperbaiki fungsi tambah agar setiap gambar yang di upload memiliki file name berdasarkan inputan name

+++ app/route/route.php
@@ -1,10 +1,6 @@
 <?php
 include '../aplikasi_bnsp/app/confiq/koneksi.php';
 
-// if (count($_GET) == 0) { // count = menghitung array
-//     include 'view/dashboard.php'; // dashboard
-// }
-
 $view = isset($_GET['view']) ? $_GET['view'] : 'dashboard';
 
 if ($view == 'dashboard') {
@@ -16,18 +12,31 @@ if ($view == 'dashboard') {
 } elseif ($view == 'editDataSiswa') {
     include '../aplikasi_bnsp/app/view/siswa/edit_siswa.php';
 } elseif ($view == 'hapusDataSiswa') {
-    // query hapus data
     $id = $_GET['id'];
-    $sqlDelete = "DELETE FROM siswa WHERE id=$id";
-    if ($koneksi->query($sqlDelete) === TRUE) {
-        echo "
+    $sqlCekSiswa = "SELECT * FROM siswa WHERE id='$id'";
+    if ($koneksi->query($sqlCekSiswa)->num_rows > 0) {
+
+        $data = $koneksi->query($sqlCekSiswa);
+        $siswa = $data->fetch_assoc();
+        unlink('../aplikasi_bnsp/assets/gambar/' . $siswa['foto']);
+
+        // query hapus data
+        $sqlDelete = "DELETE FROM siswa WHERE id=$id";
+        if ($koneksi->query($sqlDelete) === TRUE) {
+            echo "
         <script>
             alert('Data Berhasil Dihapus')
             window.location.href = '?view=dataSiswa'
         </script>";
-        // echo "Berhasil hapus data";
+        } else {
+            echo "Gagal hapus data: " . $koneksi->error;
+        }
     } else {
-        echo "Gagal hapus data: " . $koneksi->error;
+        echo "
+        <script>
+            alert('Data gagal dihapus, id tidak ditemukan')
+            window.location.href = '?view=dataSiswa'
+        </script>";
     }
 } elseif ($view == 'exportPdf') {
     include '../aplikasi_bnsp/app/view/siswa/exportpdf.php';
